Add the bitrev_vector function.
This is a fairly straightforward function. Switch elements in an
array according to the given bit-reversal table.
Unfortunately, I decided to fight with Rust's type system while
implementing the tests. In addition, the function was so simple to
reimplement that I didn't pay enough attention and actually do the last
 step of the element switching. Oops.
I guess it's good the test was there to point out my silly mistake.