Add the reduce module to ntt
After moving the reduce module into the ntt subdirectory, I need to
add it as a mod in ntt.rs. Otherwise I won't get far with
mul_coefficients.

Also, I added a simple assertion checking the lengths of vectors in
bitrev_vector. This is necessary because I just passed in the whole
dang bitrev_table instead of goofing with NEWHOPE_N and such.
Using #define macros like NEWHOPE_N *totally* works, and this is the
reference implementation after all. I couldn't write this from scratch.

Hopefully I can simplify the plumbing in such a way that it could be
upstreamed!